streamlit/requirements.txt
pandas==2.0.3
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.6
mysql-connector-python==8.2.0
